// src/interfaces/subCategoryProductsInterfave.ts
export interface SubCategoryProductsRes {
  results: number
  metadata: Metadata
  data: Data[]
}

export interface Metadata {
  currentPage: number
  numberOfPages: number
  limit: number
}

export interface Data {
  sold: number
  images: string[]
  subcategory: Subcategory[]
  ratingsQuantity: number
  _id: string
  title: string
  slug: string
  description: string
  quantity: number
  price: number
  priceAfterDiscount: number
  imageCover: string
  category: Category
  brand: Brand
  ratingsAverage: number
  createdAt: string
  updatedAt: string
  id: string
}

export interface Subcategory {
  _id: string
  name: string
  slug: string
  category: string
}

export interface Category {
  _id: string
  name: string
  slug: string
  image: string
}

export interface Brand {
  _id: string
  name: string
  slug: string
  image: string
}